+++ utils.py
@@ -1,0 +1,135 @@
+categories = {
+    'action_words': [
+        "bite", "blow", "break", "bring", "build", "bump", "buy", "carry", "catch", "chase",
+        "clap", "clean", "climb", "close", "cook", "cover", "cry", "cut", "dance", "draw",
+        "drink", "drive", "drop", "dry", "dump", "eat", "fall", "feed", "find", "finish",
+        "fit", "fix", "get", "give", "go", "have", "hear", "help", "hide", "hit",
+        "hold", "hug", "hurry", "jump", "kick", "kiss", "knock", "lick", "like", "listen",
+        "look", "love", "make", "open", "paint", "pick", "play", "pour", "pretend", "pull",
+        "push", "put", "read", "ride", "rip", "run", "say", "see", "shake", "share",
+        "show", "sing", "sit", "skate", "sleep", "slide", "smile", "spill", "splash", "stand",
+        "stay", "stop", "sweep", "swim", "swing", "take", "talk", "taste", "tear", "think",
+        "throw", "tickle", "touch", "wait", "wake", "walk", "wash", "watch", "wipe", "wish",
+        "work", "write"
+    ],
+    'animal_sounds': [
+        "baa baa", "meow", "uh oh", "choo choo", "moo", "vroom", "cockadoodledoo", "ouch", "woof woof",
+        "grr", "quack quack", "yum yum"
+    ],
+    'animals': [
+        "alligator", "animal", "ant", "bear", "bee", "bird", "bug", "bunny", "butterfly", "cat",
+        "chicken", "cow", "deer", "dog", "donkey", "duck", "elephant", "fish", "frog", "giraffe",
+        "goose", "hen", "horse", "kitty", "lamb", "lion", "monkey", "moose", "mouse", "owl",
+        "penguin", "pig", "pony", "puppy", "rooster", "sheep", "squirrel", "teddybear", "tiger",
+        "turkey", "turtle", "wolf", "zebra"
+    ],
+
+    'vehicles':[
+        "airplane", "bicycle", "boat", "bus", "car", "firetruck", "helicopter", "motorcycle", "sled",
+        "stroller", "tractor", "train", "tricycle", "truck"
+    ],
+
+    'toys' : [
+        "ball", "balloon", "bat", "block", "book", "bubbles", "chalk", "crayon", "doll", "game",
+        "glue", "pen", "pencil", "play dough", "present", "puzzle", "story", "toy"
+    ],
+
+    'food_and_drink' : [
+        "apple", "applesauce", "banana", "beans", "bread", "butter", "cake", "candy", "carrots", "cereal",
+        "cheerios", "cheese", "chicken", "chocolate", "coffee", "coke", "cookie", "corn", "cracker", "donut",
+        "drink", "egg", "fish", "food", "french fries", "grapes", "green beans", "gum", "hamburger", "ice",
+        "ice cream", "jello", "jelly", "juice", "lollipop", "meat", "melon", "milk", "muffin", "noodles",
+        "nuts", "orange", "pancake", "peanut butter", "peas", "pickle", "pizza", "popcorn", "popsicle", "potato",
+        "potato chip", "pretzel", "pudding", "pumpkin", "raisin", "salt", "sandwich", "sauce", "soda/pop", "soup",
+        "spaghetti", "strawberry", "toast", "tuna", "vanilla", "vitamins", "water", "yogurt"
+    ],
+
+    'clothing' : [
+        "beads", "belt", "bib", "boots", "button", "coat", "diaper", "dress", "gloves", "hat",
+        "jacket", "jeans", "mittens", "necklace", "pajamas", "pants", "scarf", "shirt", "shoe",
+        "shorts", "slipper", "sneaker", "snowsuit", "sock", "sweater", "tights", "underpants", "zipper"
+    ],
+
+    'body_parts' : [
+        "ankle", "arm", "belly button", "buttocks/bottom", "cheek", "chin", "ear", "eye", "face", "feet",
+        "finger", "hair", "hand", "head", "knee", "leg", "lips", "mouth", "nose", "owie/boo boo", "penis",
+        "shoulder", "tooth", "toe", "tongue", "tummy", "vagina"
+    ],
+
+    'small_household_items' : [
+        "basket", "blanket", "bottle", "box", "bowl", "broom", "brush", "bucket", "camera", "can",
+        "clock", "comb", "cup", "dish", "fork", "garbage", "glass", "glasses", "hammer", "jar",
+        "keys", "knife", "lamp", "light", "medicine", "money", "mop", "nail", "napkin", "paper",
+        "penny", "picture", "pillow", "plate", "plant", "purse", "radio", "scissors", "soap", "spoon",
+        "tape", "telephone", "tissue/kleenex", "toothbrush", "towel", "trash", "tray", "vacuum", "walker", "watch"
+    ],
+
+    'furniture_and_rooms' : [
+        "basement", "bathroom", "bathtub", "bed", "bedroom", "bench", "chair", "closet", "couch", "crib",
+        "door", "drawer", "dryer", "garage", "high chair", "kitchen", "living room", "oven", "play pen",
+        "porch", "potty", "refrigerator", "rocking chair", "room", "shower", "sink", "sofa", "stairs", "stove",
+        "table", "TV", "washing machine", "window"
+    ],
+
+    'outside_things' : [
+        "backyard", "cloud", "flag", "flower", "garden", "grass", "hose", "ladder", "lawn mower", "moon",
+        "pool", "rain", "rock", "roof", "sandbox", "shovel", "sidewalk", "sky", "slide", "snow",
+        "snowman", "sprinkler", "star", "stick", "stone", "street", "sun", "swing", "tree", "water", "wind"
+    ],
+
+    'places_to_go' : [
+        "beach", "camping", "church", "circus", "country", "downtown", "farm", "gas station", "home", "house",
+        "movie", "outside", "park", "party", "picnic", "playground", "school", "store", "woods", "work", "yard", "zoo"
+    ],
+
+    'people' : [
+        "aunt", "baby", "babysitter", "babysitter's name", "boy", "brother", "child", "child's own name", "clown", "cowboy",
+        "daddy", "doctor", "fireman", "friend", "girl", "grandma", "grandpa", "lady", "mailman", "man", "mommy", "nurse",
+        "people", "person", "pet's name", "police", "sister", "teacher", "uncle"
+    ],
+
+    'games_and_routine' : [
+        "bath", "breakfast", "bye", "call (on phone)", "dinner", "give me five!", "gonna get you!", "go potty", "hello", "hi",
+        "lunch", "nap", "night night", "no", "patty cake", "peekaboo", "please", "shopping", "shh/shush/hush", "snack",
+        "so big!", "thank you", "this little piggy", "turn around", "yes"
+    ]
+}
+
+bins = [0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 200, 300, 500, 1000]  # Extend this list if needed
+
+def calculate_month(filename):
+    weeks = int(filename[:3])
+    months = weeks // 4
+    return months
+
+# Function to map a month to its interval group
+def get_month_group(month, month_intervals):
+    for group, interval in month_intervals.items():
+        if month in interval:
+            return group
+    return None
+
+# Define the month intervals for grouping
+month_intervals = {
+    3: [0, 1, 2, 3],
+    6: [4, 5, 6],
+    9: [7, 8, 9],
+    12: [10, 11, 12],
+    18: [13, 14, 15, 16, 17, 18],
+    24: [19, 20, 21, 22, 23, 24],
+    30: [25, 26, 27, 28, 29, 30]
+}
+
+def categorize_into_bins(differences, bins):
+    bin_counts = {f"{bins[i]}-{bins[i+1]}": 0 for i in range(len(bins)-1)}
+    bin_counts[f"{bins[-1]}+"] = 0
+    for diff in differences:
+        placed = False
+        for i in range(len(bins) - 1):
+            if bins[i] <= diff < bins[i + 1]:
+                bin_counts[f"{bins[i]}-{bins[i+1]}"] += 1
+                placed = True
+                break
+        if not placed:
+            bin_counts[f"{bins[-1]}+"] += 1
+    return bin_counts
